pyproject.toml
[tool.poetry]
name = "pixel_kpi"
version = "0.1.0"
description = ""
authors = ["Jochen Luithardt <luithardt.jo96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pixoo = "^0.9.2"
python-dotenv = "^1.0.1"
numpy = "^2.1.1"
pillow = "^10.4.0"
matplotlib = "^3.9.2"
pypistats = "^1.8.0"

[tool.poetry.scripts]
pixel_kpi = "pixel_kpi.__main__:main"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
